+++ tests/testthat/test-plot_continuity.R
@@ -1,4 +1,7 @@
-# Pomocna funkce, ktera vykresluje graf do docasneho PDF (aby se nezobrazovaly)
+# --------------------------------------------------------------------------- #
+# Pomocna funkce, ktera vykresluje grafy do docasneho PDF (aby se nezobrazovaly)
+# --------------------------------------------------------------------------- #
+
 quiet_plot <- function(fun) {
   tf <- tempfile(fileext = ".pdf")
   grDevices::pdf(tf)
@@ -13,7 +16,9 @@ quiet_plot <- function(fun) {
 # --------------------------------------------------------------------------- #
 
 df <- lpanda::sample_data;
-nd <- suppressMessages(suppressWarnings(prepare_network_data(df, verbose = FALSE)));
+nd <- suppressMessages(suppressWarnings(prepare_network_data(df,
+                                                             quick = TRUE,
+                                                             verbose = FALSE)));
 
 # --------------------------------------------------------------------------- #
 
@@ -30,6 +35,8 @@ test_that("it accepts list as input data", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, do_not_print_to_console = TRUE)));
 })
 
+# --------------------------------------------------------------------------- #
+
 test_that("it fails if you put a non-existing range of elections", {
   expect_error(quiet_plot(function() plot_continuity(nd, elections = "-13", do_not_print_to_console = TRUE)));
   expect_no_error(quiet_plot(function() plot_continuity(nd, elections = "13-15", do_not_print_to_console = TRUE)));
@@ -39,21 +46,33 @@ test_that("it fails if you put a non-existing range of elections", {
   expect_error(quiet_plot(function() plot_continuity(nd, elections = "2002-", do_not_print_to_console = TRUE)));
 })
 
+# --------------------------------------------------------------------------- #
+
 test_that("links accepts only allowed values", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, links = "continuity",
                                                         do_not_print_to_console = TRUE)));
   expect_no_error(quiet_plot(function() plot_continuity(nd, links = "all",
                                                         do_not_print_to_console = TRUE)));
-  expect_error(plot_continuity(nd, links = "nope"))
+  
+  # suppressWarnings() je tak kvuli ulozeni starych grafickych parametru:
+  # "Warning in graphics::par(old_par) : calling par(new=TRUE) with no plot"
+  expect_error(suppressWarnings(plot_continuity(nd, links = "nope")));
 })
+
+# --------------------------------------------------------------------------- #
 
 test_that("lists accepts only allowed values", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, lists = "all",
                                                         do_not_print_to_console = TRUE)));
   expect_no_error(quiet_plot(function() plot_continuity(nd, lists = "elected",
                                                         do_not_print_to_console = TRUE)));
-  expect_error(plot_continuity(nd, lists = "nope"));
+  
+  # suppressWarnings() je tak kvuli ulozeni starych grafickych parametru:
+  # "Warning in graphics::par(old_par) : calling par(new=TRUE) with no plot"
+  expect_error(suppressWarnings(plot_continuity(nd, lists = "nope")));
 })
+
+# --------------------------------------------------------------------------- #
 
 test_that("parties does not error if not existed party IDs are entered", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, parties = 1000,
@@ -61,6 +80,8 @@ test_that("parties does not error if not existed party IDs are entered", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, parties = c(1,2,5),
                                                         do_not_print_to_console = TRUE)));
 })
+
+# --------------------------------------------------------------------------- #
 
 test_that("order_lists_by and order_groups_by validate values and skip those that are not in data", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, order_lists_by = "votes",
@@ -71,10 +92,16 @@ test_that("order_lists_by and order_groups_by validate values and skip those tha
   expect_no_error(quiet_plot(function() plot_continuity(nd, separate_groups = TRUE,
                                                         order_groups_by = "none",
                                                         do_not_print_to_console = TRUE)));
-  expect_error(plot_continuity(nd, order_lists_by = "nope", do_not_print_to_console = TRUE));
-  expect_error(plot_continuity(nd, separate_groups = TRUE, order_groups_by = "nope",
-                               do_not_print_to_console = TRUE));
+  
+  # suppressWarnings() je tak kvuli ulozeni starych grafickych parametru:
+  # "Warning in graphics::par(old_par) : calling par(new=TRUE) with no plot"
+  expect_error(suppressWarnings(plot_continuity(nd, order_lists_by = "nope",
+                                                do_not_print_to_console = TRUE)));
+  expect_error(suppressWarnings(plot_continuity(nd, separate_groups = TRUE, order_groups_by = "nope",
+                                                do_not_print_to_console = TRUE)));
 })
+
+# --------------------------------------------------------------------------- #
 
 test_that("mark='parties' runs and that invalid group_colour length produce warning", {
   expect_no_error(quiet_plot(function() plot_continuity(nd, mark = "parties",
@@ -83,6 +110,8 @@ test_that("mark='parties' runs and that invalid group_colour length produce warn
   expect_warning(quiet_plot(function() plot_continuity(nd, mark = "parties", group_colours = c("red"),
                                                        do_not_print_to_console = TRUE)));
 })
+
+# --------------------------------------------------------------------------- #
 
 test_that("candidate highlight does not error if candidate is not specified or does not exist", {
   
@@ -93,4 +122,71 @@ test_that("candidate highlight does not error if candidate is not specified or d
                                                         do_not_print_to_console = TRUE)));
   expect_no_error(quiet_plot(function() plot_continuity(nd, mark = c("candidate", "Kevin the Minion"),
                                                         do_not_print_to_console = TRUE)));
+})
+
+# --------------------------------------------------------------------------- #
+
+test_that("show_candidate_networks works with both data frame and network list input", {
+  
+  expect_no_error(quiet_plot(function() plot_continuity(df, show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+  
+  expect_no_error(quiet_plot(function() plot_continuity(nd, show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+})
+
+test_that("show_candidate_networks respects elections selection", {
+  
+  expect_no_error(quiet_plot(function() plot_continuity(nd, elections = "18-",
+                                                        show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+  
+  # nenavazujici volby + show_elections_between = TRUE
+  expect_no_error(quiet_plot(function() plot_continuity(nd, elections = "14, 22",
+                                                        show_candidate_networks = TRUE,
+                                                        show_elections_between = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+  
+  # nenavazujici volby + show_elections_between = FALSE
+  expect_no_error(quiet_plot(function() plot_continuity(nd, elections = c(14, 22),
+                                                        links = "all",
+                                                        show_elections_between = FALSE,
+                                                        show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+})
+
+test_that("show_candidate_networks works also with only single election", {
+  
+  expect_no_error(quiet_plot(function() plot_continuity(nd, elections = 18,
+                                                        show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+  
+  df_one_election <- df[df$elections == 18,];
+  nd_one <- suppressMessages(suppressWarnings(prepare_network_data(df_one_election,
+                                                                   skip_groups = TRUE,
+                                                                   verbose = FALSE)));
+  
+  expect_no_error(quiet_plot(function() plot_continuity(nd_one, show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+})
+
+test_that("show_candidate_networks works with party and candidate highlighting", {
+  
+  # parties + snapshots
+  expect_no_error(quiet_plot(function() plot_continuity(nd, mark = c("parties", 1),
+                                                        show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+  
+  # candidate + snapshots
+  expect_no_error(quiet_plot(function() plot_continuity(nd, mark = c("candidate", "c03"),
+                                                        show_candidate_networks = TRUE,
+                                                        do_not_print_to_console = TRUE)));
+})
+
+test_that("show_candidate_networks must be a logical scalar", {
+  expect_error(plot_continuity(nd, show_candidate_networks = "jasne",
+                               do_not_print_to_console = TRUE));
+  
+  expect_error(plot_continuity(nd, show_candidate_networks = c(TRUE, FALSE),
+                               do_not_print_to_console = TRUE));
 })